🗃Refactor: js에서 생성되는 html을 컴포넌트로 수정
