+++ packages/joy/src/constant/index.js
@@ -1,6 +1,7 @@
 //URL
 export const MAIN_URL = 'http://localhost:5510';
 export const BASE_URL = 'http://localhost:3000/api';
+export const RANDOM_IMG_URL = 'https://randomfox.ca/images';
 
 //정규식
 export const EMAIL_FORMAT_CHECK =
@@ -9,3 +10,7 @@ export const EMAIL_FORMAT_CHECK =
 //에러 메세지
 export const ERR_CHECK_PASSWORD_MSG = '패스워드를 확인해주세요.';
 export const ERR_EMAIL_FORMAT_MSG = '옳지 않은 이메일 형식입니다.';
+
+//상수
+export const RANDOM_IMG_MAX = 123;
+export const RENDER_IMG_MAX = 10;
