+++ packages/joy/src/main.js
@@ -1,13 +1,30 @@
 import '../assets/index.css';
 import { addBookmark, getBookmarkList } from './api';
 import { $, toggleLoading, debounce } from './util';
+import { heartBtnWrapper, explore_tab, saved_tab } from './components';
+import { RANDOM_IMG_URL, RANDOM_IMG_MAX, RENDER_IMG_MAX } from './constant';
 
+//무한 스크롤
+const loadMore = debounce(() => {
+  const container = $('.container');
+  const pinList = [];
+  for (let i = RENDER_IMG_MAX; i > 0; i--) {
+    pinList.push(createPin(++globalIndex));
+  }
+  container.append(...pinList);
+}, 500);
+
+//IIFE, 즉시 실행 함수
 (() => {
   const isLogin = localStorage.getItem('user_token');
-  if (isLogin !== null) return;
-
+  if (isLogin !== null) {
+    loadMore();
+    return;
+  }
   location.replace('./login.html');
 })();
+
+////////////////////////////////////////////////////////
 
 let globalIndex = 0;
 
@@ -15,17 +32,12 @@ let globalIndex = 0;
 const createPin = () => {
   toggleLoading();
   const pin = document.createElement('div');
-  const buttonWrapper = document.createElement('div');
   const image = document.createElement('img');
-  const random = Math.floor(Math.random() * 123) + 1;
-  image.src = `https://randomfox.ca/images/${random}.jpg`;
-  buttonWrapper.setAttribute('class', 'button-wrapper');
-  buttonWrapper.innerHTML = `
-  <div class="anim-icon anim-icon-md heart">
-    <input type="checkbox" id="heart${globalIndex}" />
-    <label for="heart${globalIndex}" key=${random}></label>
-  </div>
-  `;
+
+  const random = Math.floor(Math.random() * RANDOM_IMG_MAX) + 1;
+  image.src = `${RANDOM_IMG_URL}/${random}.jpg`;
+  const buttonWrapper = heartBtnWrapper(globalIndex, random);
+  
   pin.classList.add('pin');
   pin.appendChild(buttonWrapper);
   pin.appendChild(image);
@@ -33,18 +45,7 @@ const createPin = () => {
   return pin;
 };
 
-//무한 스크롤
-const loadMore = debounce(() => {
-  const container = $('.container');
-  const pinList = [];
-  for (let i = 10; i > 0; i--) {
-    pinList.push(createPin(++globalIndex));
-  }
-  container.append(...pinList);
-}, 500);
-
-loadMore();
-
+//스크롤 시 새로운 사진 추가
 window.addEventListener('scroll', () => {
   const loader = $('.loader');
   if (loader === null) return;
@@ -52,6 +53,7 @@ window.addEventListener('scroll', () => {
   loadMore();
 });
 
+//탭 전환
 $('nav').addEventListener('click', async (event) => {
   event.stopPropagation(); //이벤트 버블링 제거
   console.log(event.target);
@@ -63,41 +65,23 @@ $('nav').addEventListener('click', async (event) => {
   //선택한 탭이 explore라면
   if (event.target.matches('#explore')) {
     $main.classList.remove('saved');
-    $main.innerHTML = `
-      <div class="container"></div>
-      <div class="loader"></div>
-    `;
-
+    $main.innerHTML = explore_tab();
     globalIndex = 0;
     loadMore();
   }
 
+  //선택한 탭이 saved(=bookmark)라면
   if (event.target.matches('#saved')) {
     $main.classList.add('saved');
     const _id = localStorage.getItem('user_token');
     const result = await getBookmarkList({ _id });
-    const $content = `
-    <div class="container">
-    ${result
-      .map(
-        ({ _id, url }, index) => `
-      <div class="pin">
-        <div class="button-wrapper">
-          <div class="anim-icon anim-icon-md heart">
-            <input type="checkbox" id="heart${index}" checked>
-            <label for="heart${index}" key=${_id}></label>
-          </div>
-        </div><img src="https://randomfox.ca/images/${url}.jpg">
-      </div>`
-      )
-      .join('')}
-    </div>
-    `;
+    const $content = saved_tab(result);
 
     $main.innerHTML = $content;
   }
 });
 
+//북마크 선택
 $('main').addEventListener('click', async (event) => {
   if (!event.target.matches('label[for^="heart"]')) return;
   const _id = localStorage.getItem('user_token');
